ui: layout polish, accessibility, memoization, centered dashboard news, ticker-only top movers, modal centering
