+++ src/index.css
@@ -15,7 +15,7 @@ img { max-width: 100%; display: block; }
 button { font: inherit; cursor: pointer; }
 input, select { font: inherit; }
 
-::root {
+:root {
   --bg: #ffffff;
   --surface: #ffffff;
   --surface-alt: #f3f4f6;
@@ -25,6 +25,21 @@ input, select { font: inherit; }
   --primary: #111827;
   --success: #059669;
   --danger: #dc2626;
+  --header-h: 56px;
+}
+
+@media (prefers-color-scheme: dark) {
+  :root {
+    --bg: #0b1220;
+    --surface: #0f172a;
+    --surface-alt: #111827;
+    --text: #e5e7eb;
+    --muted: #9ca3af;
+    --border: #1f2937;
+    --primary: #e5e7eb;
+    --success: #34d399;
+    --danger: #f87171;
+  }
 }
 
 .app {
@@ -35,25 +50,43 @@ input, select { font: inherit; }
   grid-template-rows: auto 1fr;
 }
 
-.container { max-width: 1320px; margin: 0 auto; padding: 20px; }
+/* Ensure homepage content fills the viewport so the modal centers correctly */
+.main { min-height: calc(100vh -  var(--header-h)); display: flex; flex-direction: column; }
+.main > .container:last-child { margin-bottom: 0; }
+
+.container { max-width: 1200px; margin: 0 auto; padding: 16px; }
 .header {
   position: sticky; top: 0; z-index: 10;
   background: var(--surface);
   border-bottom: 1px solid var(--border);
 }
-.header-inner { display: flex; align-items: center; justify-content: space-between; gap: 12px; padding: 14px 20px; }
-.brand { font-weight: 700; letter-spacing: -0.02em; font-size: 20px; }
-.tabs { display: flex; gap: 10px; flex-wrap: wrap; }
-.tab { padding: 10px 14px; border: 1px solid var(--border); background: var(--surface); border-radius: 10px; }
+.header-inner { display: flex; align-items: center; justify-content: space-between; gap: 12px; padding: 12px 16px; min-height: var(--header-h); }
+.brand { font-weight: 700; letter-spacing: -0.02em; font-size: 18px; }
+.tabs { display: flex; gap: 8px; flex-wrap: wrap; }
+.tab { padding: 10px 14px; border: 1px solid var(--border); background: var(--surface); border-radius: 999px; transition: background .2s ease; }
 .tab.active { background: var(--surface-alt); }
 .tab:hover { background: #f8fafc; }
 
-.grid { display: grid; gap: 18px; }
+.grid { display: grid; gap: 16px; }
 .grid.cols-3 { grid-template-columns: repeat(3, minmax(0, 1fr)); }
 .grid.cols-4 { grid-template-columns: repeat(4, minmax(0, 1fr)); }
 @media (max-width: 1024px) { .grid.cols-4 { grid-template-columns: repeat(3, 1fr); } }
 @media (max-width: 768px) { .grid.cols-4, .grid.cols-3 { grid-template-columns: repeat(2, 1fr); } }
 @media (max-width: 520px) { .grid.cols-4, .grid.cols-3 { grid-template-columns: 1fr; } }
+
+/* Centered dashboard news grid */
+.news-grid-dashboard {
+  display: grid;
+  grid-template-columns: repeat(3, 1fr);
+  gap: 24px;
+  max-width: 1100px;
+  margin: 0 auto;
+  justify-content: center;
+  justify-items: center;
+}
+.news-grid-dashboard > .news-card-dashboard { width: 100%; max-width: 320px; }
+@media (max-width: 1100px) { .news-grid-dashboard { grid-template-columns: repeat(2, 1fr); max-width: 740px; } }
+@media (max-width: 740px) { .news-grid-dashboard { grid-template-columns: 1fr; max-width: 360px; } }
 
 .card { background: var(--surface); border: 1px solid var(--border); border-radius: 14px; padding: 18px; transition: box-shadow .2s ease, transform .2s ease; }
 .card:hover { box-shadow: 0 6px 24px rgba(0,0,0,.06); transform: translateY(-1px); }
@@ -61,6 +94,13 @@ input, select { font: inherit; }
 .muted { color: var(--muted); }
 .row { display: flex; align-items: center; gap: 12px; }
 .space-y-2 > * + * { margin-top: 8px; }
+
+/* Truncation/overflow helpers */
+.min-w-0 { min-width: 0; }
+.truncate { overflow: hidden; text-overflow: ellipsis; white-space: nowrap; }
+
+/* Focus visibility */
+*:focus-visible { outline: 2px solid var(--primary); outline-offset: 2px; }
 
 .stat { display: grid; grid-template-columns: 1fr auto; gap: 8px; align-items: end; }
 .chip { display: inline-flex; align-items: center; gap: 6px; padding: 2px 8px; border-radius: 999px; border: 1px solid var(--border); background: var(--surface-alt); font-size: 12px; }
@@ -72,6 +112,7 @@ input, select { font: inherit; }
 .table th { font-weight: 600; color: var(--muted); }
 .table tbody tr { transition: background .15s ease; }
 .table tbody tr:hover { background: #fafafa; }
+.table-wrap { width: 100%; overflow-x: auto; }
 
 .input { width: 100%; padding: 10px 12px; border-radius: 10px; border: 1px solid var(--border); background: var(--surface); }
 .btn { padding: 10px 12px; border-radius: 10px; border: 1px solid var(--border); background: var(--surface); }
@@ -107,9 +148,11 @@ input, select { font: inherit; }
 /* Mini coin card layout */
 .mini-card { display: grid; gap: 8px; }
 .mini-header { display: flex; align-items: center; gap: 10px; }
-.mini-title { line-height: 1.1; }
+.mini-header img { flex-shrink: 0; }
+.mini-title { line-height: 1.1; flex: 1; min-width: 0; }
+.mini-title.min-w-0 { min-width: 0; }
 .mini-symbol { font-size: 12px; }
-.mini-price { white-space: nowrap; }
+.mini-price { white-space: nowrap; flex: 0 0 auto; }
 .spark { width: 100%; height: 36px; }
 
 /* Modal */
@@ -122,3 +165,8 @@ input, select { font: inherit; }
 .media-lg { height: 180px; }
 .news-card { display: grid; gap: 8px; grid-auto-rows: min-content; }
 .news-meta { font-size: 12px; color: var(--muted); }
+.news-card-dashboard { grid-template-rows: auto auto auto; padding: 12px; gap: 6px; }
+.news-card-dashboard .card-title { font-size: 14px; font-weight: 600; margin: 4px 0 0; }
+.news-card-dashboard .news-meta { font-size: 11px; }
+.news-card-dashboard .media { border-radius: 8px; height: 96px; }
+.news-card-dashboard:hover { box-shadow: 0 8px 20px rgba(0,0,0,.06); transform: translateY(-1px); }
