+++ src/services/api.js
@@ -1,64 +1,111 @@
 // --- API CONFIGURATION ---
 export const COINGECKO_API_URL = 'https://api.coingecko.com/api/v3';
 export const CRYPTOCOMPARE_API_URL = 'https://min-api.cryptocompare.com/data/v2';
+const CRYPTOCOMPARE_API_KEY = process.env.REACT_APP_CRYPTOCOMPARE_KEY;
+
+// --- Lightweight caching + dedup + retry wrapper ---
+const cacheStore = new Map(); // key -> { expiry: number, data: any }
+const inflight = new Map(); // key -> Promise<any>
+
+const sleep = (ms) => new Promise((r) => setTimeout(r, ms));
+
+async function fetchWithRetry(url, fetchInit, retries = 2) {
+  let attempt = 0;
+  let lastErr;
+  while (attempt <= retries) {
+    try {
+      const res = await fetch(url, fetchInit);
+      if (!res.ok) {
+        // Surface Retry-After for 429
+        if (res.status === 429 && attempt < retries) {
+          const retryAfter = parseInt(res.headers.get('Retry-After') || '0', 10);
+          const backoff = (retryAfter ? retryAfter * 1000 : 500 * Math.pow(2, attempt)) + Math.floor(Math.random() * 250);
+          await sleep(backoff);
+          attempt += 1;
+          continue;
+        }
+        // Retry on 5xx
+        if (res.status >= 500 && attempt < retries) {
+          const backoff = 500 * Math.pow(2, attempt) + Math.floor(Math.random() * 250);
+          await sleep(backoff);
+          attempt += 1;
+          continue;
+        }
+        const text = await res.text().catch(() => '');
+        throw new Error(`HTTP ${res.status}: ${text || 'Request failed'}`);
+      }
+      return res.json();
+    } catch (err) {
+      lastErr = err;
+      if (attempt >= retries) break;
+      const backoff = 400 * Math.pow(2, attempt) + Math.floor(Math.random() * 250);
+      await sleep(backoff);
+      attempt += 1;
+    }
+  }
+  throw lastErr || new Error('Network error');
+}
+
+async function fetchJson(url, { ttlMs = 0, retries = 2, headers = {}, signal } = {}) {
+  const key = url;
+  const now = Date.now();
+  const cached = cacheStore.get(key);
+  if (cached && cached.expiry > now) return cached.data;
+  if (inflight.has(key)) return inflight.get(key);
+
+  const fetchInit = {
+    method: 'GET',
+    headers: { Accept: 'application/json', ...headers },
+    signal,
+  };
+  const p = fetchWithRetry(url, fetchInit, retries)
+    .then((data) => {
+      if (ttlMs > 0) cacheStore.set(key, { expiry: now + ttlMs, data });
+      inflight.delete(key);
+      return data;
+    })
+    .catch((e) => {
+      inflight.delete(key);
+      throw e;
+    });
+  inflight.set(key, p);
+  return p;
+}
 
 // --- API SERVICE FUNCTIONS ---
 
 export const fetchGlobalStats = async () => {
-  const response = await fetch(`${COINGECKO_API_URL}/global`);
-  if (!response.ok) throw new Error('Failed to fetch global stats');
-  return response.json();
+  return fetchJson(`${COINGECKO_API_URL}/global`, { ttlMs: 60_000 });
 };
 
 export const fetchTrendingCoins = async () => {
-  const response = await fetch(`${COINGECKO_API_URL}/coins/markets?vs_currency=usd&order=price_change_percentage_24h_desc&per_page=20&page=1&sparkline=false`);
-  if (!response.ok) throw new Error('Failed to fetch trending coins');
-  return response.json();
+  return fetchJson(`${COINGECKO_API_URL}/coins/markets?vs_currency=usd&order=price_change_percentage_24h_desc&per_page=20&page=1&sparkline=false`, { ttlMs: 60_000 });
 };
 
 export const fetchNews = async () => {
-  const response = await fetch(`${CRYPTOCOMPARE_API_URL}/news/?lang=EN`);
-  if (!response.ok) throw new Error('Failed to fetch news');
-  return response.json();
+  const keyParam = CRYPTOCOMPARE_API_KEY ? `&api_key=${encodeURIComponent(CRYPTOCOMPARE_API_KEY)}` : '';
+  return fetchJson(`${CRYPTOCOMPARE_API_URL}/news/?lang=EN${keyParam}`, { ttlMs: 120_000, retries: 3 });
 };
 
 export const fetchCoins = async (page = 1, perPage = 50) => {
-  const response = await fetch(`${COINGECKO_API_URL}/coins/markets?vs_currency=usd&order=market_cap_desc&per_page=${perPage}&page=${page}&sparkline=true&price_change_percentage=24h`);
-  if (!response.ok) throw new Error('Failed to fetch coins');
-  return response.json();
+  return fetchJson(`${COINGECKO_API_URL}/coins/markets?vs_currency=usd&order=market_cap_desc&per_page=${perPage}&page=${page}&sparkline=true&price_change_percentage=24h`, { ttlMs: 30_000 });
 };
 
 export const fetchCoinDetails = async (coinId) => {
-  const response = await fetch(`${COINGECKO_API_URL}/coins/${coinId}`);
-  if (!response.ok) throw new Error('Failed to fetch coin details');
-  return response.json();
+  return fetchJson(`${COINGECKO_API_URL}/coins/${encodeURIComponent(coinId)}`, { ttlMs: 60_000 });
 };
 
 export const fetchCoinHistory = async (coinId, days, interval) => {
   const safeId = encodeURIComponent(coinId);
   const safeDays = days === 'max' ? 'max' : encodeURIComponent(String(days));
   const base = `${COINGECKO_API_URL}/coins/${safeId}/market_chart?vs_currency=usd&days=${safeDays}`;
-
-  const tryFetch = async (url) => {
-    const res = await fetch(url);
-    if (!res.ok) throw new Error(`History fetch failed: ${res.status}`);
-    return res.json();
-  };
-
-  // Prefer daily when requested; fall back to default cadence on any failure
-  if (interval === 'daily') {
-    try {
-      return await tryFetch(`${base}&interval=daily`);
-    } catch (_) {
-      // fallback without interval
-      return await tryFetch(base);
-    }
-  }
-  // No interval requested
+  const url = interval === 'daily' ? `${base}&interval=daily` : base;
+  const ttl = days === '1' ? 30_000 : days === '7' ? 90_000 : days === '30' ? 180_000 : days === '90' ? 240_000 : days === '180' ? 300_000 : days === '365' ? 480_000 : 600_000;
+  // Use caching + retry and fall back to alternate cadence if first attempt fails
   try {
-    return await tryFetch(base);
+    return await fetchJson(url, { ttlMs: ttl, retries: 2 });
   } catch (_) {
-    // last resort: try daily
-    return await tryFetch(`${base}&interval=daily`);
+    const fallbackUrl = interval === 'daily' ? base : `${base}&interval=daily`;
+    return await fetchJson(fallbackUrl, { ttlMs: ttl, retries: 2 });
   }
 };
